feat(loadouts): move modlist and startup sections into own components, rename loadoutform -> createLoadoutForm
